model_eval/requirements.txt
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
flatbuffers==25.12.19
idna==3.11
ImageIO==2.37.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.46.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
onnxruntime==1.24.1
packaging==26.0
pillow==12.1.1
platformdirs==4.5.1
pooch==1.9.0
protobuf==6.33.5
PyMatting==1.1.15
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.0
sympy==1.14.0
tifffile==2026.1.28
tqdm==4.67.3
urllib3==2.6.3
